MATLAB
- CWT of a sinusoidal signal contaminated with white noise
Description: Generate a signal consisting of two sinusoids with frequencies of 100 Hz and 50 Hz contaminated
with white noise with a standard deviation of 0.25. The 100 Hz sinusoid begins at t=0 s and has a
duration of 1 s, while the 50-Hz sinusoid begins at t=3 s and has a duration of 2 s. Assume a sampling
frequency of 1000 Hz.
Perform a CWT analysis with the complex-valued Morlet wavelet, cmor1-1. Obtain a scalogram
assuming that the frequency region of interest ranges from 10 to 125 Hz and plot it as a scaled
image. This should show the percentage of energy for each wavelet coefficient.

- CWT of a fractal signal
Description: Wavelet analysis can detect fractal signals (signals showing self-similarity at different scales). Load a
Koch curve by typing load vonkoch. This curve is a synthetic fractal curve that is built recursively.
Plot the first 510 samples of the Koch curve and check visually that this is indeed a fractal curve.